requirements.txt
streamlit>=1.36
pandas>=2.1
python-dotenv>=1.0
opencv-python-headless>=4.8
Pillow>=9.4
openpyxl>=3.1
filelock>=3.12
# Optionnel pour QR dynamique (en-tête)
qrcode>=7.4
